requirements.txt
pymupdf

transformers
sentence-transformers
torch
numpy
nltk
tiktoken
protobuf

faiss-cpu

fastapi
uvicorn

black
pylint
